requirements.txt
anyio==4.8.0
bibtexparser==1.4.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
feedparser==6.0.11
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dotenv==1.0.1
pytz==2025.1
pyzotero==1.6.9
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sgmllib3k==1.0.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
